pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thesis"
version = "0.1.0"
description = "De-Americanization thesis pipeline"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "openpyxl>=3.1",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
    "databento>=0.30",
]

[tool.setuptools.packages.find]
where = ["src"]
